Criando script para classes e arrumando o main
